* merge condition-builder and use only condition
