+++ grafana/alert/index.js
@@ -2,10 +2,8 @@
 
 var Alert = require('./alert');
 var Condition = require('./condition');
-var ConditionBuilder = require('./condition-builder');
 
 module.exports = {
   Alert: Alert,
   Condition: Condition,
-  ConditionBuilder: ConditionBuilder
 };
