+++ test/alert/alert.js
@@ -1,22 +1,26 @@
-var test = require('tape');
-var Alert = require('../../grafana/alert/alert');
-var Condition = require('../../grafana/alert/condition');
-var simpleAlert = require('../fixtures/alert/simple_alert');
-var alertWithCondition = require('../fixtures/alert/alert_with_condition');
-var condition = require('../fixtures/alert/simple_condition');
+const test = require('tape');
+const Alert = require('../../grafana/alert/alert');
+const Condition = require('../../grafana/alert/condition');
+const simpleAlert = require('../fixtures/alert/simple_alert');
+const alertWithCondition = require('../fixtures/alert/alert_with_condition');
 
 test('simple alert', function t(assert) {
-  var alert = new Alert();
+  const alert = new Alert();
 
   assert.deepEqual(alert.generate(), simpleAlert);
   assert.end();
 });
 
-test('alert with conditions override', function t(assert) {
-    var alert = new Alert({
-    conditions: [condition]
-  });
+test('alert should be able to add condition', assert => {
+  const alert = new Alert();
+  const condition = new Condition()
+    .onQuery('A')
+    .withReducer('min')
+    .withEvaluator(1.1, 'gt');
+
+  alert.addCondition(condition);
 
   assert.deepEqual(alert.generate(), alertWithCondition);
   assert.end();
 });
+
