+++ test/alert/index.js
@@ -1,3 +1,2 @@
 require('./alert');
 require('./condition');
-require('./condition-builder');
